mufg_scrape/requirements.txt
beautifulsoup4==4.11.1
# chromedriver-binary==108.0.5359.22.0
chromedriver-binary==107.0.5304.62
google-cloud-secret-manager==2.12.6
Flask==2.2.2
gunicorn==20.1.0
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
retry==0.9.2
selenium==4.6.0
